Validate subnets for interconnet L2 networks
For interconnect and L2 networks, we require the subnets to be
partitioned per node. This is to avoid, for now, the need of a global ip
allocator running off cluster manager.

Additionally, all zones will have to add all the other zones pods as
remote LSPs on their transit switch. These remote ports need to have a
tunnel ID assigned that needs to be unique per network and the maximum
number of IDs that can be used is 32768, limited by OVN.

All the zones need to use the same tunnel ID for the same pods. This
tunnel ID will be generated from the assigned IP. This is to avoid, for
now, the need of a global tunnel ID allocator running off cluster
manager.

In practice, this limits the network space to 32768 pods and we
validate that. For IPv6, using a default partition of 64 does not make
sense as that would be much bigger so we use 120 bits which is
equivalent in size to the IPv4 one.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>