Split build-pr into a reusable workflow
The plan is to call this workflow from two other separate workflows
dealing each with unit tests and e2e tests.

The workflow has two input parameters: one to run unit tests and another
one to upload the resulting docker image as artifact.

Setup concurrency grouped by merge ref or run id so that the workflow
only runs once per PR or on master.

The resulting docker image is cached by merge or master ref/sha so that
it is rebuilt on any change.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>