Change EgressIPReachabiltyTotalTimeout to 10s
With the previous 1 second default and the health check implementation
that was in place before the health check controller, an unreachable
node would be declared unreachable after 7 to 12 seconds of actually
becoming unreachable.

Change the default to 10 seconds, so that with the new implementation a
node is declared unreachable after 10 to 11 seconds of actually
becoming unreachable in an attempt to keep the default behavior closer
to what we had before.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>