e2e: wait for taints to be removed when setting node back to ready
A test could fail when allocating pods to a node that has the
not-reachable taint. That could happen if the previous test last step
was to set that node back to Ready. This waits for the node condition to
be set back to ready as well but does not explcitily wait for the taint
to be removed which seems to happen asynchronously, potentially a change
of behavior from the last kube rebase.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>