RouteAdvertisements: fail if DisableMP is true
If MultiProtocol is disabled (not by default) then a BGP session
carries prefixes of both IPv4 and IPv6 families. Our problem is
that with an IPv4 session, FRR can incorrectly pick the
masquerade IPv6 address (instead of the real address) as next hop
for IPv6 prefixes and that won't work. Note that with a dedicated
IPv6 session that can't happen since FRR will use the same
address that was used to stablish the session. Let's
disallow the use of DisableMP for now.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>