Make EIP controllers aware of route advertisements
Make cluster manager, zone and node EIP controllers aware of route
advertisements.

An egress IP can be either local to a node (the node is directly
connected to the egress IP network through the primary or a secondary
interface) or advertised through a node.

Cluster manager EIP controller checks if an egress IP that is not local
to any node can be instead advertised on any node and if so, assigns it to
that node. For an egress IP to be advertised on a node, a route
advertisement configured to advertise egress IPs must select a network
that is serving any of the namespaces selected by the egress IP. If
different networks serve the namespaces selected by an egress IP, all
those newtworks must be configured to advertise egress IPs on at least
one common node for the egress IP to be assigned. To be able to reconcile
updates correctly, the controller listens to network and namespace events.

Zone and node controllers require minor changes to merely treat
advertised egress IPs as if they were local to the primary network
interface.

Egress IP utility methods are reorganized in util/egressip/util.go

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>